+++ src/device/registry.ts
@@ -2,3 +2,5 @@ import './b2500V1';
 import './b2500V2';
 import './venus';
 import './jupiter';
+import './ct002';
+import './mi800';