+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Next]
+
+- Fix time synchronization to use local timezone offset (#98)
+- Venus: Allow setting maximum charging power as low as 0W (#117)
+- Add support for CT002 smart meter device type HME (#116)
+- Add support for MI800 micro inverter device type HMI (#118)
+
+
 ## [1.4.3] - 2025-06-22
 
 - Jupiter: Add MPPT sensors for PV voltage, current, and power (#82)